Upload bar plotters
Upload modified bar-plotting code provided by Knut to plot AMAP emulator projected temperatures for 5 SSPs in 2030 and 2050, as well as CMIP6 results for the same year. The breakdown of contributions from each of the species is plotted as bars. The bar_plotters_2050_rtps.py file plots the results for SSP5-8.5, 2050, for the modified AMAP emulator.